:recycle: use reaction to manage redirection
