+++ src/screens/conference/betaAccess/betaAccess.listeners.js
@@ -1,11 +1,10 @@
 import { when } from 'k-ramel'
 
-import { redirectToNextUrl } from 'store/drivers/router/redirect'
 import betaAccess from 'firebase/betaAccess'
 import userCrud from 'firebase/user'
 
 export default [
-  when('@@ui/CHECK_BETA_ACCESS_KEY')(async (action, store, { router }) => {
+  when('@@ui/CHECK_BETA_ACCESS_KEY')(async (action, store) => {
     const key = action.payload
     const accessRef = await betaAccess.read(key)
 
@@ -16,7 +15,7 @@ export default [
       store.ui.beta.reset()
 
       // redirect to the next url if exists
-      redirectToNextUrl(router)
+      store.dispatch('@@router/REDIRECT_TO_NEXT_URL')
     } else {
       store.ui.beta.set({ error: 'Sorry, invalid beta access key.' })
     }
