+++ src/store/listeners/listeners.js
@@ -19,6 +19,8 @@ export default [
   when('@@krml/INIT')(app.init),
   /* router */
   when('@@router/ROUTE_FOUND')(router.onRouteChanged),
+  when('@@router/REPLACE_WITH_NEXT_URL')(router.replaceWithNextUrl),
+  when('@@router/REDIRECT_TO_NEXT_URL')(router.redirectToNextUrl),
   /* firebase actions */
   when('@@firebase/SIGNED_IN')(auth.signedIn),
   when('@@firebase/SIGNED_OUT')(auth.signedOut),
